Remove crate-wide and per-item clippy silences
Strip the three global #![allow(clippy::...)] directives from lib.rs
(too_many_arguments, type_complexity, needless_range_loop) and remove
per-item #[allow(clippy::too_many_arguments)] and enum_variant_names
suppressions throughout the codebase. Keep only unusual_byte_groupings
(RISC-V instruction encoding) and upper_case_acronyms (x86 mnemonics)
where the naming is dictated by hardware conventions.

+++ src/passes/div_by_const.rs
@@ -1837,7 +1837,7 @@ mod tests {
             } else {
                 let hi = ((x as u128 * magic) >> 64) as u64;
                 let diff = x - hi;
-                ((diff >> 1).wrapping_add(hi) >> (shift - 1)) as u64
+                (diff >> 1).wrapping_add(hi) >> (shift - 1)
             };
             assert_eq!(result, x / 10, "Failed for x={}", x);
         }
@@ -1858,7 +1858,7 @@ mod tests {
                 } else {
                     let hi = ((x as u128 * magic) >> 64) as u64;
                     let diff = x - hi;
-                    ((diff >> 1).wrapping_add(hi) >> (shift - 1)) as u64
+                    (diff >> 1).wrapping_add(hi) >> (shift - 1)
                 };
                 assert_eq!(result, x / d, "Failed for x={} d={}", x, d);
             }
@@ -1884,7 +1884,7 @@ mod tests {
                     } else {
                         let hi = ((x as u128 * magic) >> 64) as u64;
                         let diff = x - hi;
-                        ((diff >> 1).wrapping_add(hi) >> (shift - 1)) as u64
+                        (diff >> 1).wrapping_add(hi) >> (shift - 1)
                     };
                     assert_eq!(result, x / d, "Failed for x={} d={}", x, d);
                 }
